test/fixtures/wpt/eventsource/shared-worker/eventsource-close.htm

<!doctype html>
<html>
  <head>
    <title>shared worker - EventSource: close()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();
      test.step(function() {
        var worker = new SharedWorker('eventsource-close.js')
        worker.port.onmessage = function(e) {
          test.step(function() {
            assert_true(e.data[0], e.data[1])
            assert_equals(e.data[1], EventSource.CLOSED, 'this.readyState')
          })
          test.done()
        }
      })
    </script>
  </body>
</html>
